refactor: collectible preparation
