+++ packages/mask/src/plugins/Collectible/SNSAdaptor/OpenSea/ListingByPriceCard.tsx
@@ -12,7 +12,7 @@ import { DateTimePanel } from '../../../../web3/UI/DateTimePanel.js'
 import { toAsset } from '../../helpers.js'
 import getUnixTime from 'date-fns/getUnixTime'
 import { first } from 'lodash-unified'
-import { isWyvernSchemaName } from '../../utils.js'
+import { isWyvernSchemaName } from '../../helpers.js'
 import { useAccount, useChainId, useFungibleTokenWatched } from '@masknet/plugin-infra/web3'
 import { useOpenSea } from '../../hooks/useOpenSea.js'
 import { ActionButtonPromise } from '../../../../extension/options-page/DashboardComponents/ActionButton.js'
