+++ packages/mask/src/plugins/Collectible/SNSAdaptor/OpenSea/SelectTokenListPanel.tsx
@@ -6,7 +6,7 @@ import Check from '@mui/icons-material/Check'
 import { FormControl, ListItemIcon, MenuItem, Typography, InputBase, InputAdornment } from '@mui/material'
 import { FungibleToken, isSameAddress, formatBalance } from '@masknet/web3-shared-base'
 import type { Web3Helper } from '@masknet/plugin-infra/web3'
-import { useI18N } from '../../../utils/index.js'
+import { useI18N } from '../../../../utils/index.js'
 
 const MIN_AMOUNT_LENGTH = 1
 const MAX_AMOUNT_LENGTH = 79
